Follow up Map level and start scene with button
In this commit I continued with the development of the map level and the user interactions with the start scene with the play button. Now by running the start-scene with the play button, you can test the interaction with this button. I gave up trying to create fog in the map Level. I added new sprites for the platform and created a new player controller for the animations of the Green Dino to be personalized for each scene. Take note